requirements.txt
pycryptodome==3.20.0
Requests==2.32.2
schedule==1.2.1
